refactor(plugins): keep provider catalog registration lightweight
Move manifest catalog builders into the existing plugin owner while retaining public SDK exports. Defer live discovery imports until the selected catalog hook runs, with a regression covering the static import boundary.
