+++ src/plugin-sdk/provider-catalog-shared.ts
@@ -1,25 +1,15 @@
 // Provider catalog helpers normalize, hash, and expose model catalogs for provider plugins.
 import { createHash } from "node:crypto";
-import { normalizeModelCatalog } from "@openclaw/model-catalog-core/model-catalog-normalize";
-import { buildModelCatalogRef } from "@openclaw/model-catalog-core/model-catalog-refs";
-import type {
-  ModelCatalogCost,
-  ModelCatalogMediaInputConfig,
-  ModelCatalogModel,
-  ModelCatalogTieredCost,
-} from "@openclaw/model-catalog-core/model-catalog-types";
 import { findNormalizedProviderKey } from "@openclaw/model-catalog-core/provider-id";
 import {
   isFutureDateTimestampMs,
   resolveExpiresAtMsFromDurationMs,
 } from "../../packages/normalization-core/src/number-coercion.js";
-import { normalizeOptionalString } from "../../packages/normalization-core/src/string-coerce.js";
 import { normalizeConfiguredProviderCatalogModelId } from "../agents/model-ref-shared.js";
 import { resolveProviderRequestCapabilities } from "../agents/provider-attribution.js";
 import type { ModelDefinitionConfig } from "../config/types.models.js";
 import type { OpenClawConfig } from "../config/types.openclaw.js";
 import { pruneMapToMaxSize } from "../infra/map-size.js";
-import type { ProviderPlugin } from "../plugins/types.js";
 import type { ModelProviderConfig } from "./provider-model-shared.js";
 
 export type {
@@ -29,9 +19,14 @@ export type {
 } from "../plugins/types.js";
 
 export {
+  buildManifestModelProviderConfig,
+  buildManifestProviderCatalogFamily,
   buildPairedProviderApiKeyCatalog,
   buildSingleProviderApiKeyCatalog,
   findCatalogTemplate,
+  readManifestProviderDefaultModelRef,
+  type ManifestProviderCatalogEntry,
+  type ManifestProviderCatalogSurface,
 } from "../plugins/provider-catalog.js";
 
 /**
@@ -117,198 +112,6 @@ export async function getCachedLiveCatalogValue<T>(params: {
  */
 export function clearLiveCatalogCacheForTests(): void {
   liveCatalogCache.clear();
-}
-
-function countRawManifestCatalogModels(catalog: unknown): number | undefined {
-  if (!catalog || typeof catalog !== "object") {
-    return undefined;
-  }
-  const models = (catalog as { models?: unknown }).models;
-  return Array.isArray(models) ? models.length : undefined;
-}
-
-/** Reads a provider's normalized manifest default as a fully qualified model ref. */
-export function readManifestProviderDefaultModelRef(
-  manifest: unknown,
-  providerId: string,
-): string | undefined {
-  const catalog = (manifest as { modelCatalog?: { providers?: Record<string, unknown> } })
-    ?.modelCatalog?.providers?.[providerId];
-  const defaultModel = normalizeOptionalString(
-    (catalog as { defaultModel?: unknown })?.defaultModel,
-  );
-  return defaultModel ? buildModelCatalogRef(providerId, defaultModel) : undefined;
-}
-
-function cloneManifestCatalogTieredCost(
-  tier: ModelCatalogTieredCost,
-): NonNullable<ModelDefinitionConfig["cost"]["tieredPricing"]>[number] {
-  return {
-    input: tier.input,
-    output: tier.output,
-    cacheRead: tier.cacheRead,
-    cacheWrite: tier.cacheWrite,
-    range: tier.range.length === 1 ? [tier.range[0]] : [tier.range[0], tier.range[1]],
-  };
-}
-
-function cloneManifestCatalogCost(cost: ModelCatalogCost): ModelDefinitionConfig["cost"] {
-  return {
-    input: cost.input ?? 0,
-    output: cost.output ?? 0,
-    cacheRead: cost.cacheRead ?? 0,
-    cacheWrite: cost.cacheWrite ?? 0,
-    ...(cost.tieredPricing
-      ? { tieredPricing: cost.tieredPricing.map(cloneManifestCatalogTieredCost) }
-      : {}),
-  };
-}
-
-function buildManifestCatalogModelInput(model: ModelCatalogModel): ModelDefinitionConfig["input"] {
-  if (model.input?.includes("document")) {
-    throw new Error(
-      `Manifest modelCatalog row ${model.id} uses unsupported runtime input document`,
-    );
-  }
-  return model.input?.filter((item): item is "text" | "image" => item !== "document") ?? ["text"];
-}
-
-function cloneManifestCatalogMediaInput(
-  mediaInput?: ModelCatalogMediaInputConfig,
-): ModelDefinitionConfig["mediaInput"] | undefined {
-  if (!mediaInput?.image) {
-    return undefined;
-  }
-  return {
-    image: { ...mediaInput.image },
-  };
-}
-
-function buildManifestCatalogModel(
-  providerId: string,
-  model: ModelCatalogModel,
-): ModelDefinitionConfig {
-  if (model.contextWindow === undefined) {
-    throw new Error(`Manifest modelCatalog row ${model.id} is missing contextWindow`);
-  }
-  if (model.maxTokens === undefined) {
-    throw new Error(`Manifest modelCatalog row ${model.id} is missing maxTokens`);
-  }
-  const id = normalizeConfiguredProviderCatalogModelId(providerId, model.id, {
-    allowManifestNormalization: false,
-  });
-  return {
-    id,
-    name: model.name ?? id,
-    ...(model.api ? { api: model.api } : {}),
-    ...(model.baseUrl ? { baseUrl: model.baseUrl } : {}),
-    reasoning: model.reasoning ?? false,
-    input: buildManifestCatalogModelInput(model),
-    cost: cloneManifestCatalogCost(model.cost ?? {}),
-    contextWindow: model.contextWindow,
-    ...(model.contextTokens !== undefined ? { contextTokens: model.contextTokens } : {}),
-    maxTokens: model.maxTokens,
-    ...(model.thinkingLevelMap ? { thinkingLevelMap: { ...model.thinkingLevelMap } } : {}),
-    ...(model.headers ? { headers: { ...model.headers } } : {}),
-    ...(model.compat ? { compat: { ...model.compat } } : {}),
-    ...(model.mediaInput ? { mediaInput: cloneManifestCatalogMediaInput(model.mediaInput) } : {}),
-  };
-}
-
-/**
- * Converts a plugin manifest modelCatalog provider into runtime provider config.
- */
-export function buildManifestModelProviderConfig(params: {
-  /** Provider id that owns the manifest catalog rows. */
-  providerId: string;
-  /** Raw manifest modelCatalog provider block to normalize into runtime config. */
-  catalog: unknown;
-}): ModelProviderConfig {
-  const catalog = normalizeModelCatalog(
-    { providers: { [params.providerId]: params.catalog } },
-    { ownedProviders: new Set([params.providerId]) },
-  )?.providers?.[params.providerId];
-  if (!catalog) {
-    throw new Error(`Missing modelCatalog.providers.${params.providerId}`);
-  }
-  if (!catalog.baseUrl) {
-    throw new Error(`Missing modelCatalog.providers.${params.providerId}.baseUrl`);
-  }
-  const rawModelCount = countRawManifestCatalogModels(params.catalog);
-  if (rawModelCount !== undefined && rawModelCount !== catalog.models.length) {
-    throw new Error(`Invalid modelCatalog.providers.${params.providerId}.models`);
-  }
-  return {
-    baseUrl: catalog.baseUrl,
-    ...(catalog.api ? { api: catalog.api } : {}),
-    ...(catalog.headers ? { headers: { ...catalog.headers } } : {}),
-    models: catalog.models.map((model) => buildManifestCatalogModel(params.providerId, model)),
-  };
-}
-
-export type ManifestProviderCatalogSurface = {
-  id: string;
-  label: string;
-  catalog: unknown;
-};
-
-export type ManifestProviderCatalogEntry = {
-  id: string;
-  label: string;
-  baseUrl: string;
-  models: ModelProviderConfig["models"];
-  buildProvider: () => ModelProviderConfig;
-};
-
-/** Projects an ordered family of manifest catalogs into static provider and model surfaces. */
-export function buildManifestProviderCatalogFamily(params: {
-  surfaces: readonly ManifestProviderCatalogSurface[];
-  docsPath?: string;
-}) {
-  const entries: ManifestProviderCatalogEntry[] = params.surfaces.map((surface) => {
-    const buildProvider = () =>
-      buildManifestModelProviderConfig({
-        providerId: surface.id,
-        catalog: surface.catalog,
-      });
-    const provider = buildProvider();
-    return {
-      id: surface.id,
-      label: surface.label,
-      baseUrl: provider.baseUrl,
-      models: provider.models,
-      buildProvider,
-    };
-  });
-  const staticDiscovery: ProviderPlugin[] = entries.map(({ id, label, buildProvider }) => ({
-    id,
-    label,
-    docsPath: params.docsPath ?? "/providers/models",
-    auth: [],
-    staticCatalog: {
-      order: "simple",
-      run: async () => ({ provider: buildProvider() }),
-    },
-  }));
-
-  return {
-    entries,
-    staticDiscovery,
-    staticCatalog: async () => ({
-      providers: Object.fromEntries(entries.map(({ id, buildProvider }) => [id, buildProvider()])),
-    }),
-    augmentModelCatalog: () =>
-      entries.flatMap(({ id: provider, models }) =>
-        models.map((entry) => ({
-          provider,
-          id: entry.id,
-          name: entry.name,
-          reasoning: entry.reasoning,
-          input: [...entry.input],
-          contextWindow: entry.contextWindow,
-        })),
-      ),
-  };
 }
 
 function normalizeConfiguredCatalogModelInput(
