+++ src/plugin-sdk/provider-entry.ts
@@ -11,6 +11,11 @@ import type {
 } from "../plugins/manifest-types.js";
 import { createProviderApiKeyAuthMethod } from "../plugins/provider-api-key-auth.js";
 import { projectProviderCatalogResultToUnifiedTextRows } from "../plugins/provider-catalog-unified-text.js";
+import {
+  buildManifestModelProviderConfig,
+  buildSingleProviderApiKeyCatalog,
+  readManifestProviderDefaultModelRef,
+} from "../plugins/provider-catalog.js";
 import type {
   ProviderPlugin,
   ProviderCatalogContext,
@@ -25,21 +30,20 @@ import {
   isRecordWithoutThrowing,
   readRecordValue,
 } from "../shared/safe-record.js";
+import { createLazyRuntimeMethod, createLazyRuntimeModule } from "./lazy-runtime.js";
 import { definePluginEntry } from "./plugin-entry.js";
 import type {
   OpenClawPluginApi,
   OpenClawPluginConfigSchema,
   OpenClawPluginDefinition,
 } from "./plugin-entry.js";
-import {
-  buildOpenAICompatibleProviderCatalog,
-  type OpenAICompatibleModelDiscoveryOptions,
-} from "./provider-catalog-live-runtime.js";
-import {
-  buildManifestModelProviderConfig,
-  buildSingleProviderApiKeyCatalog,
-  readManifestProviderDefaultModelRef,
-} from "./provider-catalog-shared.js";
+import type { OpenAICompatibleModelDiscoveryOptions } from "./provider-catalog-live-runtime.js";
+
+// Registration needs static metadata; live discovery loads only when its catalog hook runs.
+const buildOpenAICompatibleProviderCatalog = createLazyRuntimeMethod(
+  createLazyRuntimeModule(() => import("./provider-catalog-live-runtime.js")),
+  (runtime) => runtime.buildOpenAICompatibleProviderCatalog,
+);
 
 type ApiKeyAuthMethodOptions = Parameters<typeof createProviderApiKeyAuthMethod>[0];
